💀 I Really Misunderstood Classes
🔧 Classes are ACTUALLY OOP now
➕ Added EventHandler class
➕ Added GuildRoleUpdateHandler class
🗑️ Removed logChannel database utils, made it a .env field instead

The main reason for the classes change was because I discovered a bug
whilst writing the GuildRoleUpdateHandler class that was caused by a
fundamental flaw with how I'd been coding classes up until that point.
The issue was simultaneous events firing at once which caused the class
fields to be overwritten multiple times due to being within the same
class. Not an issue on its own, however because each event had to wait
for a promise to return what ended up happening was that all events got
to that point but only the LATEST event set the class fields. This meant
that EVERY event that fired simulatenous, once the promise returns, will
go on to use the field values of the latest event, breaking everything.
This is of course solved by... actually implementing classes correctly😝